Refactor stylesheet compilation into an Eleventy plugin
Using `tasks/sass.js` as base, exports an Eleventy plugin from `elevenyt/stylesheets.js`.
Within the file, extracts a shared glob for which files are watched and compiled to ensure they match
